Adds decoding of core transceiver monitors
- Restructures the `decode` crate a bit, splitting out modules by what
  they decode.
- Adds the `monitors` module, with a `Monitors` type for pulling out the
  temp, voltage and per-lane Rx power and Tx bias current and power.
- Adds SFF-8636 decoding, with test.
- Adds CMIS decoding with tests.
- Adds methods to `Controller` and commands to `xcvradm` for exercising
  and printing the monitors
